PCR brake: Init _openRealTime from first PCR, like in discont. readjust
[Committer canvon] This should avoid a potentially long wait
just after opening a file, if the first PCR happens to have
a large value-as-seconds.

It could still be a problem when the first PCR's value-as-seconds
is larger than the system's uptime or whatever the CLOCK_MONOTONIC
started from, as a negative value for _openRealTime is used as a flag.