Copy back modified Adaptation Field to TS packet
Setting the Discontinuity Indicator would not have worked before,
as it would have operated on a QByteArray that was a copy of the
relevant part (Adaptation Field) of the QByteArray in the TS packet
object.

So this should now be more correct; but in practice it doesn't really
help with stopping mpv from halting. It seems to introduce audio-visual
glitches between input files, though.