StreamServer: Open input non-blocking
This can avoid the whole program blocking on, e.g., a FIFO
until the other side (runs and) opens it.

* Introduce _inputFileName to hold the file name when it is cleared
  in the QFile due to fd-based open().

This seems to have the problem, though, that many peer processes
can write to the FIFO without the server necessarily getting an EOF
on it. (Although this might have been possible before.)