ts-split: Implement missing start/length kinds, esp. segment-based
* Rename Splitter::setOutputs() to setOutputRequests(); we now have
  both outputRequests and outputResults, the latter counting the
  length unit actually used up until now.
* Support all 3 StartKinds: startOffset, startPacket &
  startDiscontSegment.
* Support all 3 LengthKinds: lenBytes, lenPackets &
  lenDiscontSegments.
* The support is internally as well as for the --outfile command-line
  option.

A discontinuity segment can now be extracted using:

  --outfile startdiscontsegment=3,lendiscontsegments=1,filename=outfoo3.ts

N.B.: Running ts-split in this way on multiple concatenated transport
streams will not recover those transport streams exactly. This is due to
the algorithm currently cutting the input stream on first PCR, which
might not be the first packet in the stream.

E.g., consider a file test.ts consisting of parts A before the first
PCR, PCR, and B after the first PCR.

| cat test.ts test.ts test.ts >triple-test.ts
| ./ts-split --outfile ... --outfile ... --outfile ... tripe-test.ts

This will currently create the following outputs:

1 A:PCR:B:A
2 PCR:B:A
3 PCR:B

That is, file 2 will be missing part A at the front and have a trailing
part A at the end that should not be there. File 1 will have a trailing
part A at the end, too. File 3 will not contain part A at all!

This may or may not be a problem. In a test with mpv, all 3 output files
could be played back without any perceivable problems. There might still
be missing important meta-data from part A (that mpv could do without).