libinfra/HumanReadable: Refactor: Use namespace instead of nested classes
The API should stay the same, though the ABI changes(?).
(HumanReadable::Foo can both address an inner class or part of
a namespace.) At least it still seems to build and run fine. (?)

The intention is to be able to add more files that add to the
HumanReadable::Foo concept, without the need to be declared all in the
same header file.