ts-split: Refactor: Move some code to impl
(For potential later code reuse. For separating out logically
what could be perceived as logically an action of its own, too.)

Furthermore, this way, the start/open vs. finish/close operation
can be given in the source code "in logical order".

(The startOutputRequest() "that" pointer (cf. "this" pointer)
could be made unnecessary if the impl carried it around anyhow,
but I'm not going to take this step at this time.)