libmedia: TS::bslbf: Forward value construction to base
This allows initializing bslbf-s, when used in a class
as non-static data members, in various "natural" ways.
I'm told that this would not be necessary with C++17,
where it should Just Work [TM]. (The problem being that
a class with a base class is not an aggregate before C++17. (?))

Thanks go to private IRC for the help.

* Also make sure to inherit public in bslbf<1, bool> specialization,
  too.