libmedia: Add TS::Packet2Parser
TS::Packet2Parser is intended to become the replacement for
the parts of TSPacket that won't be replaced by TS::Packet2
(as the latter is mostly a value type).

In other words, we'll now have a split value<->parser setup.
Looks promising, so far, although the team of TS::Packet2 &
TS::Packet2Parser does not support AdaptationField or payload data,
yet.