libmedia: Optimize BitStream: Move implementation to header file
This reduces the example run's total cost from 10.7 million
to 8.7 million. The downside is that BitStream now no longer is
recognized by valgrind-callgrind...

This might have been a idea in any case, as the BitStream definitions
now are shorter and possibly easier to read without the artificial
"_implPtr->" or "impl." complication... Plus, the class can now be used
without the need for (own) heap allocation(; QByteArray may still use
the heap).

* Don't #include <memory> anymore, as we only needed it for
  std::unique_ptr, which we now don't use anymore.
* #include <typeinfo> which is necessary for typeid().
  It probably only worked accidentally, before; maybe <memory>
  included <typeinfo> (possibly indirectly) for us.