libmedia: Rename tsparser.h to tsprimitive.h
* Remove the .cpp which was not necessary anymore even before
  the rename.

The idea is to have something more of a "telling name".
The real parser (which makes use of this file) currently is
in tspacketv2.{h,cpp}, so that name was incorrect. Contained
in the now-renamed file are TS::BitStream and several MPEG-TS
primitive data types (and bit source operations on them),
so name them that way.

N.B.: The previous "tsparser" branch will be retired and replaced
by a new "tspacketv2" branch, accordingly.