libmedia/tsprimitive.h: Add initial code to compose source bits to a bit sink
That's the other way round than we did already; that is,
we had bit stream to MPEG-TS primitives output already (bit source),
where now we have primitives input to bit stream (bit sink).

This is not optimized (e.g. for "easy", byte-aligned cases), yet.
That makes the initial code simpler, though.

Also, there is no check against out-of-range values. Perhaps they
should be detected (especially in two's-complement cases) and
an exception should be thrown... But what about the performance
impact of that?

TODO: Add test cases.