TS::PacketV2Parser: Refactor: Replace struct Parse with in-implementation ParseState
Part of the idea is to be able to pass in the output TS::PacketV2
instance by address. That theoretically allows using a derived class,
instead. Also both a packet on the heap (for later queueing for clients)
or on the stack (as in ts-dump) can be used directly.

While at it, the error message is now optional, and if not passed in,
potential error message generation will be skipped.

libmedia/tspacketv2:
* Move basic packet parsing from public parser to implementation,
  as parsePacket(). Pass newly added ParseState as implementation
  arguments.
* Split packet size verification into wider packet verification
  in public parse(), and enough bytes being left for basic packet size
  in implementation parsePacket().

* libmedia/tsprimitive: Remove "explicit" marker from TS::BitStream
  to be able to implicitly construct it from a QByteArray in parse()...

* ts-dump: Adjust use of TS::PacketV2Parser accordingly.