TS primitives: Fix output into bit sink for uimsbf/tcimsbf larger than an int
.., e.g., for components of a PCR, with 33bit uimsbf > 32bit int...

For bitshifting a 1-bit to a mask, we first need to convert the 1 to the
mask type / working type. Otherwise, "funny" things happen with
completely wrong masks, and I've heared that's because shifting an n-bit
integer by n or more bits is undefined behaviour in C (and apparently
in C++, too?), so anything can happen ... like the "funny" masks, here.