Bugfix: Overcome misunderstanding about QObject::destroyed signal
It seems when the signal is delivered, the object can no longer be
cast down to its real type; just the QObject is left. So we have to
cast up any existing pointer to QObject to compare it to the signal
argument. (It seems to work, but let's hope this isn't a violation
of the C++ Standard again which might require much more "fixing"...)

* HTTP::Server:
  * Signal handler handleClientDestroyed():
    Instead of casting down & doing QList::removeOne(), cast up & use
    Java-style iterators, to make this easy(?) and safe (w.r.t.
    otherwise maybe conflating list indices...)
  * Signal clientDestroyed: Adjust from passing a more-specific type
    to passing the QObject base class, just as the QObject::destroyed
    signal does.
* StreamServer:
  * Signal handler handleStreamClientDestroyed():
    Drop correctness check & replace use of QList::removeOne()
    with use of upcast in the check & Java-style iterators.
  * Adjust handler for signal HTTP::Server::clientDestroyed.
    (See above.)
* StreamClient: Signal handler handleHTTPServerContextDestroyed():
  Drop the correctness check alltogether; the handler is private
  and will hopefully have been correctly wired. ...