libinfra/log_backend: Move optional prefix to after debugging aids
This way, when running with -vv -d, the function name logged
will line up with other ones regardless of whether the message
is a qDebug() message ("DEBUG: " prefix) or not (mostly no prefix).

The downside is that, in debug mode, one has to look to after
the function name so see, e.g., a debug or fatal prefix...
But in debug mode that may be okay, and outside of debug mode,
the prefix should still be the first thing after the timestamp
(and optional category, it seems).