libinfra/humanreadable: Hexdump: Stop dumping 0x"ff" for empty
Dump "(empty)", instead.

The former "0x\"ff\"" didn't mean an 0xff byte itself, but
**zero times** the all-one-bits byte, 0xff. That is, the data
to be dumped was just empty (0 bytes long), but the compression
automatics "compressed" it to 0 0xff bytes.

(This brought me to confusion and to the (wrong!) assumption,
the data in question would have consisted of an 0xff byte itself,
and assumed there would have been an attempt to initialize a buffer
to 255 bytes, but initialized it to an 255 (0xff) byte instead.
**All wrong!** The dumped data was just *empty*. So dump it that way.)