tests: Add testcase streamserver-cvn-cli/http/httpresponse
(This is just to get used to the test infrastructure.
Not much testing done, yet.)

This should run on `make check` in the build root, now.

During the build, before, it'll have snatched plain `.o` object files
from the main application directory, as the qmake `.pro` file
directed it. So testing parts of an application project (instead of
just of an internal library) should be possible, too, now.